XNS4 has the second argument for `gethostname` as an int. Perhaps this was ammended in corrigenda, but that needs commenting accordingly if so.
svn path=/trunk/; revision=3208

+++ tspec/base/cae/xns4-ar/unistd.h.ts
@@ -6,4 +6,4 @@
 # See doc/copyright/ for the full copyright terms.
 
 # CAE, Networking Services, Issue 4
-+FUNC int gethostname ( char *, size_t ) ;
++FUNC int gethostname ( char *, int ) ;
